{{cookiecutter.project_name}}/requirements.txt
hydra-core>=1.0.0,<2.0.0
mlflow>=2.0.0,<3.0.0
-r requirements_module.txt
-r requirements_tests.txt
-r requirements_app.txt